Set default worker to 1 (#1106)
KFServer uses the number of cores on the machine to set the number of process in its tornado service which can delay the initialization time and increase the memory usage a lot.

A local testing suggests that KFServer loaded 64 processes and used more than 500Mb for a simple SKLearn model. Once I set the worker number to 1, the memory usage was reduced to around 100Mb. 
```
[I 200927 18:26:06 kfserver:88] Registering model: mlguild-iris-svm
[I 200927 18:26:06 kfserver:77] Listening on port 8080
[I 200927 18:26:06 kfserver:79] Will fork 0 workers
[I 200927 18:26:06 process:126] Starting 64 processes
```