Re-add crd as base to default kustomization (#1388)
The crd installation was separated from the default install to
get around an issue with the crd size being too large to store in the
metadata.annotations.last-applied-configuration field of the
resource (max 256kb).

Currently, with the removal of initContainers spec from the inferenceservice
crd, the size is now small enough to fit into the annotations. This
commit readds the crd yaml to the default yaml.