Add env field to ServingRuntime builtInAdapter settings (#2123)
* Add env field to ServingRuntime builtInAdapter settings

Built-in runtimes that use the built-in adapter container have certain specific parameters configured via the builtInAdapter section in the CRD.

However, there are certain other properties of the built-in runtimes that we may want to set/change for various reasons, so it would be good to extend this with a map of arbitrary environment variables to use for this purpose. Having this means that we can expose such additional properties for the built-in runtimes without requiring further changes to the CRD.

Note this will currently only be applicable for the built-in modelmesh multi-model runtimes.

Signed-off-by: Nick Hill <nickhill@us.ibm.com>

* Rename struct fieldname to Env and re-generate openapi schema

Signed-off-by: Nick Hill <nickhill@us.ibm.com>