Introduce pathTemplate in ingressConfig
Extend the IngressConfig struct with a pathTemplate field to enable
path-based routing. If this is set the user must also provide
ingressDomain and urlScheme.

Validate the given template just by trying to parse it. The user is
responsible for providing a valid template that is unique for a given
inference service, for example, by using both namespace and name in the
template.

Introduce a path module (similar to the existing domain) that has
GenerateUrlPath() helper. This will will try to render the pathTemplate
based on the name and namespace of an isvc.

Add unittests for this new helper.

Refs kserve/kserve#2257

Signed-off-by: Dimitris Aragiorgis <dimara@arrikto.com>
Signed-off-by: Kostis Andriopoulos <kandrio@arrikto.com>