Add path-based rule in top-level VirtualService
Extend the Ingress reconciler to add a matching rule in the top-level
VirtualService of each InferenceService if the 'pathTemplate' field is
not empty.

Specifically adds a VirtualService rule that redirects path-based requests from
the public to the local gateway. Also extend the list of hosts to include the
ingressDomain.

Modify existing getServiceUrl() to take into account pathTemplate. Update
unittests accordingly.

Refs kserve/kserve#2257

Signed-off-by: Dimitris Aragiorgis <dimara@arrikto.com>
Signed-off-by: Kostis Andriopoulos <kandrio@arrikto.com>