python: remove last traces of log_config_file (#2851)
After a chat with @yuzisun in:

https://github.com/kserve/website/pull/266

it was suggested to remove `log_config_file` in favor
of a more standardized `log_config`. To keep argparse
and the model server's signatures inline, this change
adds the possibility to accept as input a dictionary
"serialized" as string, unpacking it with the json library.
I am not sure what users will use this functionality,
but let's discuss if we want it or not as suggested
by @yuzisun.

Signed-off-by: Luca Toscano <toscano.luca@gmail.com>