sklearnserver image: Enable running under arbitrary UID
Set file's group to the root group and give the group the same permissions as the users.
This is the recommendation of OpenShift (https://docs.openshift.com/container-platform/4.13/openshift_images/create-images.html#use-uid_create-images) to run under arbitrary user ids (but with those users also being member of the "root" group).

This change does is non-intrusive and will work on any platform.

Signed-off-by: Roland Huß <rhuss@redhat.com>