Revert and re-adapt test/e2e/common/utils.py
This reverts changes done in commit ecff079 to test/e2e/common/utils.py.
It is possible to revert since opendatahub-io/odh-model-controller#59 is
solved.

After the revert, another adaptation was needed. CI in upstream assumes
that the URL in the InferenceService cannot be resolved, so it uses the
IP of the istio-ingressgateway together with the HTTP Host header to
workaround this limitation. However, in ODH testing environment the URL
of InferenceServices can be resolved. The adaptation adds a flag to use
the URL of the isvc, rather than using the IP of the
istio-ingressgateway.

This adaptation can be contributed to upstream, because the community
can benefit from it.

The CI setup is updated, so that the revert works correctly:
* odh-model-controller is now being installed
* TLS is removed from the istio-ingressgateway

Additionally, the following other changes were done to the CI setup:
* Turn off Mesh-related components that are not needed for CI
* Prefer usage of the ServiceMeshMember CRD to enroll namespaces to the
  Mesh

Fixes opendatahub-io/kserve#50

Signed-off-by: Edgar Hernández <23639005+israel-hdez@users.noreply.github.com>