Add a CI_USE_ISVC_HOST for testing with the ISVC hostname (#3324)
* Add a CI_USE_ISVC_HOST for testing with the ISVC hostname

This adds the possibility to run E2E tests using the hostname specified in the `status.url` field of the InferenceServices, rather than using the ClusterIP of the Istio Ingress gateway.

This behavior is enabled by, first, setting the `CI_USE_ISVC_HOST=1` environment variable and, then invoking pytest. Thus, the original behavior of using Istio IngressGateway ClusterIP is the default one.

Signed-off-by: Edgar Hernández <23639005+israel-hdez@users.noreply.github.com>

* Add a CI_USE_ISVC_HOST for testing with the ISVC hostname

This adds the possibility to run E2E tests using the hostname specified in the `status.url` field of the InferenceServices, rather than using the ClusterIP of the Istio Ingress gateway.

This behavior is enabled by, first, setting the `CI_USE_ISVC_HOST=1` environment variable and, then invoking pytest. Thus, the original behavior of using Istio IngressGateway ClusterIP is the default one.

Signed-off-by: Edgar Hernández <23639005+israel-hdez@users.noreply.github.com>

---------

Signed-off-by: Edgar Hernández <23639005+israel-hdez@users.noreply.github.com>