Add compatibility for Istio CNI plugin (#3316)
When Istio is installed with its CNI plugin, KServe inference services are not capable to start. This is because the storage initializer is an init-container and the network is not available when the CNI plugin is enabled.

The typical recommendation to fix the issue is to remove init-containers and move any initialization code to a regular container. This approach would not work well with KServe, because the serving runtimes assume the model is already present on the filesystem and moving the storage initializer as a regular container will cause race conditions (the runtime will succeed loading only if the storage initializer manages to pull the model before the runtime starts).

There are alternative approaches documented in https://istio.io/latest/docs/setup/additional-setup/cni/#compatibility-with-application-init-containers. All alternatives have the downside that the traffic won't be captured by Istio and won't benefit from Istio features, which should be OK for KServe storage-initializer case.

These changes use the approach for running the storage initializer using the same UserID as the Istio sidecar. The UID is copied from the sidecar container to cover Istio derivatives, like Maistra.

Signed-off-by: Edgar Hernández <23639005+israel-hdez@users.noreply.github.com>