Fix Raw Logger E2E
This is a small fix to the test_raw_logger.py/test_kserve_logger E2E test.

This E2E test was creating a `message-dumper-raw` that was dependant on a Serverless setup, leading to the E2E to fail on an environment where Istio nor KNative are available.

The fix is to add the right annotation to `message-dumper-raw` to deploy it in Raw mode, and adjust the logger URL of the `isvc-logger-raw` to point to the Kubernetes Service, rather than the KNative service.

Signed-off-by: Edgar Hernández <23639005+israel-hdez@users.noreply.github.com>