Reduce E2Es dependency on CI environment (#3435)
Some code of the E2Es assume the environment is GitHub, because it is referring to GitHub-specific variables.

This PR focuses on references to the `kserve/image-transformer` image. This image is built in the CI flow and made available to the runner, so that a pull from an external registry is not needed.

The references to this image are changed to an environment variable that is more agnostic to the runner, in an effort to make E2Es more compatible in other environments. The e2e-test.yml is modified to declare the new variable.

Signed-off-by: Edgar Hernández <23639005+israel-hdez@users.noreply.github.com>