Enhance controller setup based on available CRDs (#3472)
* Enhance controller setup based on available CRDs

This enhances the setup of the InferenceService controller and the InferenceGraph controller. Instead of relying on the `defaultDeploymentMode` configuration to determine what CRDs to watch, the setup now checks whether KNative Services and Istio VirtualServices are available in the cluster and setup the watches (invoke `Owns`) accordingly.

This enhancement has the following advantages:
* A crashloop is prevented if the CRDs are missing in the cluster. The user would still be able to create InferenceServices by taking care of annotating the ISVC for RawDeployment mode.
* If RawDeployment mode is configured as the default mode, the controllers would still watch for KNative and Istio resources if these components are available. This will let the controller watch for changes for the dependent resources if the user uses Serverless mode for some of the InferenceServices.
* In the InferenceService controller, the watch for the VirtualServices is still conditioned to the value of the `disableVirtualHost` configuration.

Signed-off-by: Edgar Hernández <23639005+israel-hdez@users.noreply.github.com>

* Controller setup - add schemas based on CRDs available

Since KServe controllers are modified to watch resources based on
available CRDs, a similar change in the setup of the manager is needed:
schemas need to be added to the manager based on available CRDs rather
than based only on the values in the inferenceservice-config ConfigMap.
This would keep both manager setup and controller setup in sync with
regards schemas and watches around the CRDs.

Signed-off-by: Edgar Hernández <23639005+israel-hdez@users.noreply.github.com>

---------

Signed-off-by: Edgar Hernández <23639005+israel-hdez@users.noreply.github.com>
Signed-off-by: asd981256 <asd981256@gmail.com>