Allow configuring replicas in for raw deployments
This adds the ability to configure a hard number of replicas when using Raw deployment mode.

The replica count cannot be customized by default, because autoscaling needs to be taken into account. The only two supported autoscalers were the standard HPA (default), and External ones. In both cases, the `replicas` field of the deployment is assumed to be managed by a third-party controller. Thus, the user didn't have a way to manually manage the replicas of deployed models.

To let users to manually manipulate pod replicas for raw deployments, a new `none` autoscaler class is added. This new autoscaler class would indicate to KServe controller that there is no autoscaling, and will let users set a hard replica count by using a new `replicas` field in the InferenceService CRD.

This commit augments only the InferenceService CRD with the replicas field. For InferenceGraph, replicas still won't be user-controllable.

Signed-off-by: Edgar Hernández <23639005+israel-hdez@users.noreply.github.com>