unintentional changes in go.sum
The changes to `go.sum` were unintentional. They were caused by running Go tests locally, which sometimes updates the `go.sum` file to ensure dependency integrity.
I've since reverted `go.sum` to its previous state and excluded it from this PR, as these changes are not related to the main purpose of this update, which focuses on "fixing the readiness probe logic."

Signed-off-by: Snehomoy <164564725+LOADBC@users.noreply.github.com>
Signed-off-by: Snehomoy <snehomoy6@gmail.com>