Add well-known label to InferenceGraphs
We require to add an `AuthorizationPolicy` to the resulting pods of InferenceGraph (serverless mode).
In Istio, Workload Selectors require a <key, value> pair. Apparently, there is no support to match only on the presence of a label key. See https://istio.io/latest/docs/reference/config/type/workload-selector/#WorkloadSelector.

Thus, to correctly apply AuthorizationPolicies, this is adding the `serving.kserve.io/kind: InferenceGraph` to pods that belong to InferenceGraph.

Signed-off-by: Edgar Hernández <23639005+israel-hdez@users.noreply.github.com>