Use leader election for better compatibility with in-place upgrades
For in-place upgrades the manifests of the new version are applied in a cluster where KServe may already be installed. The old version of the kserve-controller is terminated and the new version is deployed.

The default strategy is Rolling. This has the consequence that during an in-place upgrade, there will be two different versions of kserve-controller running in parallel. Most importantly, the two instances will be running with different configs, which include references to different versions of kserve-agent, kserve-router, storage-initializer, etc. During an in-place upgrade, the new version of the controller will upgrade the deployed models to use the new versions of the images. However, since the old version of kserve-controller is running in parallel, this old version will do rollback. The two instances will conflict with each other, leading to Deployments/KSVCs being changed very quickly which will lead to the cluster spawning pods without control.

Eventually, the cluster will terminate the old version of kserve-controller. However, because of the very, very quick updates to deployments the resources of the cluster can be exhausted leading to instability.

This is enabling leader election in the manager, by default. The leader election solves mentioned issues because the new version of the manager won't fully boot until it is capable of getting the lease for being the leader. The manifests enable leader election by default, because `manager_auth_proxy_patch.yaml` is used by default. However, the kustomization.yaml file gives the alternative to use `manager_prometheus_metrics_patch.yaml` and leader election won't be enabled in such case.

Signed-off-by: Edgar Hernández <23639005+israel-hdez@users.noreply.github.com>