Use Headless Services by default (#523) (#528)
This fixes a regression, because in ODH the Services had been headless for Raw deployment mode (see #288). Such default changed to headed services when the feasibility to configure the behavior was introduced, but despite the upstream project has always been using headed services, in ODH this  breaks backwards compatibility.

Thus, this restores the ODH default of using headless services for Raw deployment. Users can still use the inferenceservice-config ConfigMap to switch to headed Services.

Signed-off-by: Edgar Hernández <23639005+israel-hdez@users.noreply.github.com>