InferenceGraph: Fix response code when condition step is not fulfilled
When a step in the InferenceGraph has a condition, if it is not fulfilled for a run, such step is skipped and the step is skipped. Currently, when the step is skipped, the returned HTTP status code is 500. This is not compatible with KServe's python client which raises an exception when receiving the 500 status code.

This is changing the status code to 200 to let the python client to work correctly. It also makes more sense to return 200, since there were no errors processing the request.

Signed-off-by: Edgar Hernández <23639005+israel-hdez@users.noreply.github.com>