Resolve inference endpoint using runtime protocol when applicable
Typically, protocol v1 is assumed. However, there is a corner case: if the runtime name is specified in the InferenceService, then kserve-controller won't validate that the protocol in the InferenceService and the ServingRuntime matches (likely, assuming the user knows what is doing). More importantly, if the protocol is not specified in the InferenceService, the GetPredictorEndpoint function returns the V1 endpoint even if the ServingRuntime only supports V2 (and correctly specifies it). This has an impact on the InferenceGraph feature because the kserve-router will be misconfigured.

These changes cover this corner case to properly resolve the protocol.

Signed-off-by: Edgar Hernández <23639005+israel-hdez@users.noreply.github.com>