fix(codegen): pins code-generator binaries version
Without pinning the version of code-generator cmds, the underlying script
falls back to downloading the latest vesion. This can result in regenerating
client stubs with content diverging from the one under version control
even if the current work does not touch this part of the project. That results
in noise and can pollute PR with unrelated (or even breaking) changes.

To have consistent behavior the `KUBE_CODEGEN_TAG` is set to ensure that
codegen binaries are matching defined version dependency in the project.

Fixes #4532

Signed-off-by: Bartosz Majsak <bartosz.majsak@gmail.com>