feat(envtest): simplifies CRD lookup
Relying on relative paths for resolving CRD to be loaded when starting
EnvTest is error prone.

This PR introduces a simple `GetProjectRoot()` lookup function that
traverses project tree up until it finds `go.mod` file and assumes this
is the root of the project. That is sufficient to make EnvTest tests
resilent.

Also... makes the code legal ;)

Signed-off-by: Bartosz Majsak <bartosz.majsak@gmail.com>