fix: Allow CA bundle path without config map
In the previous code, the AWS_CA_BUNDLE environment variable is only
considered if the CA_BUNDLE_CONFIGMAP_NAME environment variable is
set, while the doc doesn't explicitely tie they together. This change
allows using AWS_CA_BUNDLE independently to leverage a CA bundle
certificate that would be preexistent in the image or injected by
Kubernetes in a known path.

One use case is leveraging OpenShift service serving certificates, which
injects the service CA in the containers via a projected ConfigMap. The
service CA is always available in
/var/run/secrets/kubernetes.io/serviceaccount/service-ca.crt.

Signed-off-by: Fabien Dupont <fdupont@redhat.com>